requirements.txt
python-telegram-bot==20.7
yadisk==2.2.0
httpx==0.27.0
apscheduler==3.10.4
charset-normalizer==3.4.3
python-docx==1.1.2